feat: extract gateway_profiles package (service + schema)
Move ProfilesService and ProfileResponse into a new gateway_profiles
package. gateway_oss keeps re-exports at services/profiles.py and
models/schemas.py so downstream extensions' imports remain unchanged.
