+++ packages/gateway_oss/pyproject.toml
@@ -18,6 +18,7 @@ dependencies = [
   "gateway_following",
   "gateway_notes",
   "gateway_posts",
+  "gateway_profiles",
   "gateway_notes_rest",
   "gateway_notes_mcp",
   "gateway_following_rest",
@@ -43,6 +44,7 @@ gateway_comments = { workspace = true }
 gateway_following = { workspace = true }
 gateway_notes = { workspace = true }
 gateway_posts = { workspace = true }
+gateway_profiles = { workspace = true }
 gateway_notes_rest = { workspace = true }
 gateway_notes_mcp = { workspace = true }
 gateway_following_rest = { workspace = true }
