+++ packages/gateway_oss/src/gateway_oss/models/schemas.py
@@ -14,14 +14,12 @@ from gateway_notes.schemas import (  # noqa: F401
     NoteResponse,
     NotesPageResponse,
 )
-from gateway_oss.models.substack import (
-    SubstackPublicProfile,
-)
 from gateway_posts.schemas import (  # noqa: F401
     FullPostResponse,
     PostResponse,
     PostsPageResponse,
 )
+from gateway_profiles.schemas import ProfileResponse  # noqa: F401
 
 
 class LivenessResponse(BaseModel):
@@ -34,23 +32,3 @@ TokensInfo = BearerCredentials
 class HealthResponse(BaseModel):
     connected: bool
     tokens: TokensInfo | None = None
-
-
-class ProfileResponse(BaseModel):
-    id: int
-    handle: str
-    name: str
-    url: str
-    avatar_url: str
-    bio: str | None = None
-
-    @classmethod
-    def from_substack(cls, profile: SubstackPublicProfile) -> ProfileResponse:
-        return cls(
-            id=profile.id,
-            handle=profile.handle,
-            name=profile.name,
-            url=f"https://substack.com/@{profile.handle}",
-            avatar_url=profile.photo_url or "",
-            bio=profile.bio,
-        )
