+++ packages/gateway_profiles/pyproject.toml
@@ -1,0 +1,15 @@
+[project]
+name = "gateway_profiles"
+version = "3.1.0"
+requires-python = ">=3.11"
+dependencies = ["gateway_core", "pydantic>=2.13.4", "aiocache[redis]>=0.12.3"]
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/gateway_profiles"]
+
+[tool.uv.sources]
+gateway_core = { workspace = true }
