+++ pyproject.toml
@@ -29,6 +29,7 @@ gateway_comments = { workspace = true }
 gateway_following = { workspace = true }
 gateway_notes = { workspace = true }
 gateway_posts = { workspace = true }
+gateway_profiles = { workspace = true }
 gateway_notes_rest = { workspace = true }
 gateway_notes_mcp = { workspace = true }
 gateway_following_rest = { workspace = true }
@@ -50,7 +51,7 @@ select = ["E", "F", "I", "UP"]
 ignore = ["E501"]
 
 [tool.ruff.lint.isort]
-known-first-party = ["gateway_core", "gateway_comments", "gateway_following", "gateway_notes", "gateway_notes_rest", "gateway_notes_mcp", "gateway_following_rest", "gateway_following_mcp", "gateway_posts", "gateway_posts_rest", "gateway_posts_mcp", "gateway_comments_rest", "gateway_comments_mcp", "gateway_oss", "gateway_rest_common", "gateway_mcp_common", "substack_gateway"]
+known-first-party = ["gateway_core", "gateway_comments", "gateway_following", "gateway_notes", "gateway_notes_rest", "gateway_notes_mcp", "gateway_following_rest", "gateway_following_mcp", "gateway_posts", "gateway_posts_rest", "gateway_posts_mcp", "gateway_comments_rest", "gateway_comments_mcp", "gateway_profiles", "gateway_oss", "gateway_rest_common", "gateway_mcp_common", "substack_gateway"]
 
 [tool.ruff.format]
 quote-style = "double"
@@ -71,6 +72,7 @@ testpaths = [
   "packages/gateway_following/tests",
   "packages/gateway_notes/tests",
   "packages/gateway_posts/tests",
+  "packages/gateway_profiles/tests",
 ]
 addopts = "--import-mode=importlib"
